Small fixes responding to PR comments

+++ app/models/user.rb
@@ -930,7 +930,9 @@ class User < ActiveRecord::Base
     @raw_password = pw # still required to maintain compatibility with usage of password-related User interface
   end
 
-  def remove_password()
+  def remove_password
+    raise Discourse::InvalidAccess if associated_accounts.blank? && passkey_credential_ids.blank?
+
     user_password.destroy if user_password
   end
